+++ RELEASE_NOTES.md
@@ -20,6 +20,7 @@ Release Notes
 - `RowModel.edited` has been renamed `RowModel.databaseEdited`.
 - `RowModel.databaseTable` has been replaced with `RowModel.databaseTableName()` which returns a String.
 - `RowModel.setDatabaseValue(_:forColumn:)` has been removed. Use and override `RowModel.updateFromRow(_)` instead.
+- `RowModel.didFetch()` has been renamed - `RowModel.awakeFromFetch()`
 
 
 ## 0.11.0